chore: Remove client-state.ts
Keep all the yjs mutators in mutators.js

Cleanup exports
