--- src/client-state.ts
@@ -1,97 +1,0 @@
-// This file defines the ClientState entity that we use to track
-// cursors. It also defines some basic CRUD functions using the
-// @rocicorp/rails helper library.
-
-import type {
-  ClientID,
-  JSONObject,
-  JSONValue,
-  ReadTransaction,
-  ReadonlyJSONObject,
-  WriteTransaction,
-} from '@rocicorp/reflect';
-import {yjsAwarenessKey, yjsAwarenessPrefix} from './mutators.js';
-
-// ClientState is where we store the awareness state
-export type ClientState = JSONObject;
-
-export async function yjsSetLocalStateField(
-  tx: WriteTransaction,
-  {
-    name,
-    yjsClientID,
-    field,
-    value,
-  }: {name: string; yjsClientID: number; field: string; value: JSONValue},
-) {
-  const clientState = await getClientState(tx, name, tx.clientID, yjsClientID);
-  if (clientState) {
-    await putClientState(tx, name, tx.clientID, yjsClientID, {
-      ...(clientState as object),
-      [field]: value,
-    });
-  }
-}
-
-export async function yjsSetLocalState(
-  tx: WriteTransaction,
-  {
-    name,
-    yjsClientID,
-    yjsAwarenessState,
-  }: {
-    name: string;
-    yjsClientID: number;
-    yjsAwarenessState: ReadonlyJSONObject | null;
-  },
-) {
-  if (yjsAwarenessState === null) {
-    await deleteClientState(tx, name, tx.clientID, yjsClientID);
-  } else {
-    await putClientState(tx, name, tx.clientID, yjsClientID, yjsAwarenessState);
-  }
-}
-
-function deleteClientState(
-  tx: WriteTransaction,
-  name: string,
-  reflectClientID: ClientID,
-  yjsClientID: number,
-) {
-  return tx.del(yjsAwarenessKey(name, reflectClientID, yjsClientID));
-}
-
-function putClientState(
-  tx: WriteTransaction,
-  name: string,
-  reflectClientID: string,
-  yjsClientID: number,
-  yjsAwarenessState: ReadonlyJSONObject | null,
-) {
-  return tx.set(
-    yjsAwarenessKey(name, reflectClientID, yjsClientID),
-    yjsAwarenessState,
-  );
-}
-
-function getClientState(
-  tx: WriteTransaction,
-  name: string,
-  reflectClientID: string,
-  yjsClientID: number,
-) {
-  return tx.get(yjsAwarenessKey(name, reflectClientID, yjsClientID));
-}
-
-export async function listClientStates(
-  tx: ReadTransaction,
-  name: string,
-): Promise<[ClientID, number, JSONObject][]> {
-  const entries: [ClientID, number, JSONObject][] = [];
-  const prefix = yjsAwarenessPrefix(name);
-  for await (const [key, value] of tx.scan({prefix}).entries()) {
-    const [reflectClientID, yjsClientID] = key.slice(prefix.length).split('/');
-    entries.push([reflectClientID, Number(yjsClientID), value as JSONObject]);
-  }
-  return entries;
-}
