+++ src/index.tsx
@@ -8,7 +8,8 @@ import {CodemirrorBinding} from 'y-codemirror';
 import * as Y from 'yjs';
 import './index.css';
 import styles from './index.module.css';
-import {Mutators, Provider, mutators} from './provider.js';
+import {Mutators, mutators} from './mutators.js';
+import {Provider} from './provider.js';
 import {UserInfo, randUserInfo} from './user-info.js';
 
 const userID = nanoid();
