+++ src/monaco.ts
@@ -10,7 +10,8 @@ import {nanoid} from 'nanoid';
 import {MonacoBinding} from 'y-monaco';
 import * as Y from 'yjs';
 import './index.css';
-import {Provider, mutators as yjsMutators} from './provider.js';
+import {mutators as yjsMutators} from './mutators.js';
+import {Provider} from './provider.js';
 import {randUserInfo} from './user-info.js';
 
 const userID = nanoid();
