+++ src/mutators.ts
@@ -16,10 +16,15 @@
 // thing. The Reflect sync protocol ensures that the server-side result takes
 // precedence over the client-side optimistic result.
 
-import type {ClientID, WriteTransaction} from '@rocicorp/reflect';
+import type {
+  ClientID,
+  ReadTransaction,
+  ReadonlyJSONObject,
+  ReadonlyJSONValue,
+  WriteTransaction,
+} from '@rocicorp/reflect';
 import * as base64 from 'base64-js';
 import * as Y from 'yjs';
-import {yjsSetLocalState, yjsSetLocalStateField} from './client-state.js';
 
 export const mutators = {
   yjsSetLocalStateField,
@@ -27,19 +32,7 @@ export const mutators = {
   updateYJS,
 };
 
-export type M = typeof mutators;
-
-export type UpdateYJS = {
-  updateYJS: typeof updateYJS;
-};
-
-export type YJSSetLocalStateField = {
-  yjsSetLocalStateField: typeof yjsSetLocalStateField;
-};
-
-export type YJSSetLocalState = {
-  yjsSetLocalState: typeof yjsSetLocalState;
-};
+export type Mutators = typeof mutators;
 
 export async function updateYJS(
   tx: WriteTransaction,
@@ -55,7 +48,7 @@ export async function updateYJS(
   }
 }
 
-export function yjsAwarenessPrefix(name: string) {
+function yjsAwarenessPrefix(name: string) {
   return `yjs/awareness/${name}/`;
 }
 
@@ -69,4 +62,94 @@ export function yjsAwarenessKey(
   yjsClientID: number,
 ): string {
   return `${yjsAwarenessPrefix(name)}${reflectClientID}/${yjsClientID}`;
+}
+
+export async function yjsSetLocalStateField(
+  tx: WriteTransaction,
+  {
+    name,
+    yjsClientID,
+    field,
+    value,
+  }: {
+    name: string;
+    yjsClientID: number;
+    field: string;
+    value: ReadonlyJSONValue;
+  },
+) {
+  const clientState = await getClientState(tx, name, tx.clientID, yjsClientID);
+  if (clientState) {
+    await putClientState(tx, name, tx.clientID, yjsClientID, {
+      ...(clientState as object),
+      [field]: value,
+    });
+  }
+}
+
+export async function yjsSetLocalState(
+  tx: WriteTransaction,
+  {
+    name,
+    yjsClientID,
+    yjsAwarenessState,
+  }: {
+    name: string;
+    yjsClientID: number;
+    yjsAwarenessState: ReadonlyJSONObject | null;
+  },
+) {
+  if (yjsAwarenessState === null) {
+    await deleteClientState(tx, name, tx.clientID, yjsClientID);
+  } else {
+    await putClientState(tx, name, tx.clientID, yjsClientID, yjsAwarenessState);
+  }
+}
+
+function deleteClientState(
+  tx: WriteTransaction,
+  name: string,
+  reflectClientID: ClientID,
+  yjsClientID: number,
+) {
+  return tx.del(yjsAwarenessKey(name, reflectClientID, yjsClientID));
+}
+
+function putClientState(
+  tx: WriteTransaction,
+  name: string,
+  reflectClientID: string,
+  yjsClientID: number,
+  yjsAwarenessState: ReadonlyJSONObject | null,
+) {
+  return tx.set(
+    yjsAwarenessKey(name, reflectClientID, yjsClientID),
+    yjsAwarenessState,
+  );
+}
+
+function getClientState(
+  tx: WriteTransaction,
+  name: string,
+  reflectClientID: string,
+  yjsClientID: number,
+) {
+  return tx.get(yjsAwarenessKey(name, reflectClientID, yjsClientID));
+}
+
+export async function listClientStates(
+  tx: ReadTransaction,
+  name: string,
+): Promise<[ClientID, number, ReadonlyJSONObject][]> {
+  const entries: [ClientID, number, ReadonlyJSONObject][] = [];
+  const prefix = yjsAwarenessPrefix(name);
+  for await (const [key, value] of tx.scan({prefix}).entries()) {
+    const [reflectClientID, yjsClientID] = key.slice(prefix.length).split('/');
+    entries.push([
+      reflectClientID,
+      Number(yjsClientID),
+      value as ReadonlyJSONObject,
+    ]);
+  }
+  return entries;
 }
